pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "live-note"
version = "0.1.0"
description = "Audio note-taking for meetings, lectures, and recordings with whisper.cpp and Obsidian"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "opencc-python-reimplemented>=0.1.7",
  "sounddevice>=0.4.7",
  "webrtcvad-wheels>=2.0.14",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0",
  "ruff>=0.12.0",
]

[project.scripts]
live-note = "live_note.app.cli:main"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
